+++ spec/support/academies_api_helpers.rb
@@ -74,4 +74,12 @@ module AcademiesApiHelpers
     allow(test_client).to receive(:get_trust).with(ukprn).and_raise(Api::AcademiesApi::Client::Error)
     allow(Api::AcademiesApi::Client).to receive(:new).and_return(test_client)
   end
+
+  def mock_unauthorised_api_responses(urn:, ukprn:)
+    test_client = Api::AcademiesApi::Client.new
+
+    allow(test_client).to receive(:get_trust).with(ukprn).and_raise(Api::AcademiesApi::Client::UnauthorisedError)
+    allow(test_client).to receive(:get_establishment).with(ukprn).and_raise(Api::AcademiesApi::Client::UnauthorisedError)
+    allow(Api::AcademiesApi::Client).to receive(:new).and_return(test_client)
+  end
 end